refactor: 구조 단순화, 필요하지 않은 interface 제거
