+++ src/main/java/racingcar/AppConfig.java
@@ -1,17 +1,13 @@
 package racingcar;
 
 import racingcar.controller.GameController;
-import racingcar.controller.GameControllerImpl;
 import racingcar.service.GameService;
-import racingcar.service.GameServiceImpl;
 import racingcar.service.picker.NumberPicker;
 import racingcar.service.picker.RandomNumberPicker;
 import racingcar.view.input.InputView;
-import racingcar.view.input.InputViewImpl;
 import racingcar.view.input.reader.ConsoleReader;
 import racingcar.view.input.reader.Reader;
 import racingcar.view.output.OutputView;
-import racingcar.view.output.OutputViewImpl;
 
 public class AppConfig {
 
@@ -38,7 +34,7 @@ public class AppConfig {
     }
 
     private OutputView initOutputView() {
-        return new OutputViewImpl();
+        return new OutputView();
     }
 
     private Reader initReader() {
@@ -46,7 +42,7 @@ public class AppConfig {
     }
 
     private InputView initInputView(Reader reader) {
-        return new InputViewImpl(reader);
+        return new InputView(reader);
     }
 
     private NumberPicker initNumberPicker() {
@@ -54,11 +50,11 @@ public class AppConfig {
     }
 
     private GameService initGameService(NumberPicker numberPicker) {
-        return new GameServiceImpl(numberPicker);
+        return new GameService(numberPicker);
     }
 
     private GameController initGameController(GameService gameService, InputView inputView, OutputView outputView) {
-        return new GameControllerImpl(gameService, inputView, outputView);
+        return new GameController(gameService, inputView, outputView);
     }
 
 }
